correct ramp_enable channel type

+++ src/dodal/devices/temperture_controller/lakeshore/lakeshore_io.py
@@ -26,7 +26,7 @@ class LakeshoreControlChannel(Device):
 
         self.user_setpoint = channel_rw(channel_type=float, pv_name="SETP")
         self.ramp_rate = channel_rw(channel_type=float, pv_name="RAMP")
-        self.ramp_enable = channel_rw(channel_type=float, pv_name="RAMPST")
+        self.ramp_enable = channel_rw(channel_type=int, pv_name="RAMPST")
         self.heater_output = channel_rw(channel_type=float, pv_name="HTR")
         self.heater_output_range = channel_rw(channel_type=heater_type, pv_name="RANGE")
         self.p = channel_rw(float, "P")
